requirements.txt
google-auth
requests
slack-webhook
